refactor: frontend code, especially DisplayForecast(s)
