+++ src/pages/recursion.tsx
@@ -1,6 +1,7 @@
+import { NextPage } from "next";
 import React, { useEffect } from "react";
 
-function Recursion() {
+const Recursion: NextPage = () => {
   useEffect(() => {
     if (typeof window !== "undefined") {
       window.location.href = "https://www.youtube.com/watch?v=dQw4w9WgXcQ";
@@ -12,6 +13,6 @@ function Recursion() {
       <h2>You have now reached the fourth level of recursion!!</h2>
     </div>
   );
-}
+};
 
 export default Recursion;
