+++ src/web/display/MultiSelectPlatform.tsx
